packages/data_foundation/pyproject.toml
[project]
name = "data_foundation"
description = "Code location for dagster that exposes dbt, sling, and dlt resources for Analytics Engineering"
version = "0.1.0"
readme = "README.md"
authors = [{name = "Andrew Staus"}]
requires-python = ">=3.13, <3.14"
dependencies = [
    "dagster-cloud==1.11.15",
    "dagster-dbt==0.27.15",
    "dagster-dlt==0.27.15",
    "dagster-postgres==0.27.15",
    "dagster-sling==0.27.15",
    "dagster-snowflake==0.27.15",
    "psutil==7.1.0",
    "data-platform-utils",
    "scikit-learn",
    "pandas",
    "pyarrow"
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]

[tool.uv.sources]
data-platform-utils = { path = "../../libs/data_platform_utils" }

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "data_foundation.definitions"
code_location_name = "data_foundation"

[tool.dg]
directory_type = "project"
project.root_module = "data_foundation"
